Set Up Project Structure
This main folder should have two folders in it: /server and /client, which, as their name tells, one will have the files related to the server (API) and the other one will have the files related to the client (UI).

Server: Initialize the project using npm init in the server folder to create a Node.js project. Next, install the required dependencies for this basic PERN Stack example using npm:

express: to use the Express web framework in the project;
pg: to use PostgreSQL in the project;
dotenv: to securely store and manage environment variables in the project, while separating them from code.

Client: Initialize the project using npx create-react-app client. After that, we need to install the required dependencies, this time for the front-end:

axios: to enable making HTTP requests from the front-end.
Co-Authored-By: Nicole Wu <143554135+Nicc041@users.noreply.github.com>
Co-Authored-By: Tutoria-Team <230203310+Tutoria-Team@users.noreply.github.com>